finish test cases for Scientific class

+++ src/main/java/com/zipcodewilmington/scientificcalculator/Basic.java
@@ -2,19 +2,19 @@ package com.zipcodewilmington.scientificcalculator;
 
 public class Basic {
 
-    public int add(int x, int y) {
+    public double add(double x, double y) {
         return x + y;
     }
 
-    public int multiply(int x, int y) {
+    public double multiply(double x, double y) {
         return x * y;
     }
 
-    public int divide(int x, int y) {
+    public double divide(double x, double y) {
         return x / y;
     }
 
-    public int subtract(int x, int y) {
+    public double subtract(double x, double y) {
         return x - y;
     }
 }
